# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UMKM Go API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List stores (admin or own)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated stores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreListResponse"
    post:
      operationId: createStore
      tags: [stores]
      summary: Create a new store (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreRequest"
      responses:
        "201":
          description: Store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stores/my:
    get:
      operationId: getMyStore
      tags: [stores]
      summary: Get current user's store
      responses:
        "200":
          description: User's store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "404":
          description: No store found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stores/{slug}:
    get:
      operationId: getStoreBySlug
      tags: [stores]
      summary: Get public store by slug (storefront)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public store with products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicStore"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateStore
      tags: [stores]
      summary: Update store
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoreRequest"
      responses:
        "200":
          description: Updated store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products for current user's store
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Store:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        categoryId:
          type: string
        categoryName:
          type: string
        whatsapp:
          type: string
        logoUrl:
          type: string
        theme:
          type: string
          enum: [modern, classic]
        userId:
          type: string
        status:
          type: string
          enum: [active, inactive]
        productCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - categoryId
        - userId
        - theme
        - status
        - productCount
        - createdAt
    PublicStore:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        categoryName:
          type: string
        whatsapp:
          type: string
        logoUrl:
          type: string
        theme:
          type: string
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
      required:
        - id
        - name
        - slug
        - products
    CreateStoreRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        categoryId:
          type: string
        whatsapp:
          type: string
        logoUrl:
          type: string
        theme:
          type: string
          enum: [modern, classic]
      required:
        - name
        - slug
        - categoryId
    UpdateStoreRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string
        whatsapp:
          type: string
        logoUrl:
          type: string
        theme:
          type: string
          enum: [modern, classic]
    StoreListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Store"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    Product:
      type: object
      properties:
        id:
          type: string
        storeId:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        isAvailable:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - storeId
        - name
        - price
        - isAvailable
        - sortOrder
        - createdAt
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        isAvailable:
          type: boolean
      required:
        - name
        - price
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        isAvailable:
          type: boolean
        sortOrder:
          type: integer
    ProductListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
...