# lib/api-zod/src/generated/types/storeListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UMKM Go API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Store } from "./store";

export interface StoreListResponse {
  data: Store[];
  total: number;
  page: number;
  limit: number;
}