# lib/api-zod/src/generated/types/updateStoreRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UMKM Go API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateStoreRequestTheme } from "./updateStoreRequestTheme";

export interface UpdateStoreRequest {
  name?: string;
  description?: string;
  categoryId?: string;
  whatsapp?: string;
  logoUrl?: string;
  theme?: UpdateStoreRequestTheme;
}